Fixes 2 more tasks in #464 - regarding viewport correct sizing on client changes
The new fixes attempt to fix an effect of over optimization. We'd previously been trying to limit the amount of time we reset the viewport. That had caused a bug where we missed when a non-active tab was in an old larger size prior to a smaller client connect, but as we only had a single 'per browser' viewport memory, and we'd already update the viewport of the active tab, when switching back to this larger tab, after the browser viewport had already been updated, our test showed there was no change to be made. This is because we didn't have a per target viewport memory. Now we do. This fixes the bug.