BrowserBox moves to PolyForm Non-commercial License with commercial options and abandons AGPL due to shady practices by non-compliant customers
Unfortunately a non-insignificant number of customers have recently engaged in shady practices, tempted by the deceptive illusion of open source licenses being a 'legal gray area', and the lure of potentially obtaining something for free. This has included: failure to disclose commercial use, failure to purchase licenses for closed-source modifications used commercially. These are backed up by tactics such as becoming non-responsive, and saying, 'we were considering open sourcing in future, and didn't think we needed a license until we had decided for sure.' which sounds reasonable but is an abuse of the letter, and the spirit, of open source licenses like the AGPL. It's also a greedy attempt to get, and use, IP for free. While we have benefitted from a number of large, compliant customers who have invested significantly in licenses and been forthcoming, transparent and honest about their use, a non-significant number of smaller, but still well known companies have engaged in deceptive and violatory tactics when trying to use BrowserBox under AGPL. This is unfortunate, as the dream of an open source utopia which can power business growth of OSS vendors, while contributing to the ecosystem, was a nice idea. But I think in some cases, such as here, where software is much more product focused, and less of a plug-in-able utility or function, OSS licensing can create the wrong incentives, and tempt otherwise not-evil companies into illegal activity.
Almost as if OSS provides the wrong expectation that 'therefore it should be free', and some companies in the current climate then spend quite a bit of time trying to figure out ways to avoid paying for licenses even when they're in breach of contract. Almost as if, once the OSS label is applied, some companies cannot stomach the idea of paying vendors for their IP because they are under the delusion that 'it should therefore be free'.
This is inefficient, for us at DOSYAGO, as we waste time dealing with these customers and entertaining their fanciful pretensions of licensing which turn out to be merely false pretenses to get some free support. But I think it's also inefficient for those customers as instead of simply paying the fee and moving on with business, they spend time crafting, employing and perfecting their deception. Time they could have spent shipping, dealing with their own customers, etc. Plus they need to then hold the added worry-debt that steals their mind share that one day they will be caught out. Well, that day has come.
So it is with some wistful tears, but stern resolve, we wave goodbye to AGPL-land as we set sail for brighter climes