Audio works in new updates except if you try to connect to same browser from tor browser and regular browser at same time, while in regular site mode.
If in hidden service mode, audio works in tor browser. if in regular mode audio works in reguilar browser. this is an edge case while annoying, is not a priotiy to fix immediately. iow the only case where it doesn't work is connecting to regular hostname (non onion) using tor browser. likley because of two things: 1) we are resetting / borking the audio server with two competing data types, tor requires wave and the other is a special chunked format. 2) we may not be detecting, in the regular site version of browserbox that we need to render it for a tor browser, so we may be still trying to utilize audiocontex. this 2nd is an easier problem. resolve the logic and correct it to point at the already working case when we are in that situation. the first issue (potential issue) of having the audio service handle fan out on the same audio source into potentially two different formats, may be more work to do, as i don't think we have considered or scaffolded for that case before.